[doc] Add a doc describing undocumented/secret API methods
There are some secret API methods in glitch.  Some are used by the
flash app, some are used by the internal docs tool, and some are
used internally on the website.

Some of them require a special 'site' role, that you can't get via
the oauth APIs and are used by the site internally.